requirement.txt
langgraph
notebook
pydantic
langchain-core
langchain-community
langchain-groq
pymupdf
sqlite3
faker